// src/app/models/episode.model.ts
// src/app/models/episode.model.ts
export interface Episode {
  id: number;
  name: string;
  air_date: string;
  episode: string;
  url: string;
  created: string;
}

export interface EpisodeResponse {
  info: {
    count: number;
    pages: number;
    next: string | null;
    prev: string | null;
  };
  results: Episode[];
}